refactor: moved out multi value
